requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
courlan==0.9.5
dateparser==1.2.0
decorator==5.1.1
distro==1.9.0
filelock==3.13.1
fsspec==2023.12.2
google==3.0.0
h11==0.14.0
htmldate==1.7.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
jusText==3.0.0
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openai==1.10.0
packaging==23.2
py==1.11.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
retry==0.9.2
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sympy==1.12
tld==0.13
tokenizers==0.15.1
torch==2.1.2
tqdm==4.66.1
trafilatura==1.7.0
transformers==4.37.1
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
